requirements.txt
# RAG Framework and Orchestration
llama-index>=0.12.0
llama-index-llms-ollama
llama-index-embeddings-huggingface
llama-index-vector-stores-chroma
llama-index-retrievers-bm25

# Vector Database
chromadb>=0.5.0

# Telegram Bot Interface
aiogram>=3.4.0

# Environment and Utilities
python-dotenv
logging